+++ src/main/java/insane96mcp/survivalreimagined/data/generator/SRGlobalLootModifierProvider.java
@@ -1,9 +1,9 @@
 package insane96mcp.survivalreimagined.data.generator;
 
-import insane96mcp.survivalreimagined.module.experience.feature.EnchantmentsFeature;
 import insane96mcp.survivalreimagined.module.farming.feature.Crops;
 import insane96mcp.survivalreimagined.module.items.feature.FlintExpansion;
 import insane96mcp.survivalreimagined.module.items.feature.RecallIdol;
+import insane96mcp.survivalreimagined.module.mining.feature.Lapis;
 import insane96mcp.survivalreimagined.module.mobs.feature.Equipment;
 import insane96mcp.survivalreimagined.module.movement.feature.Minecarts;
 import insane96mcp.survivalreimagined.module.world.feature.CoalFire;
@@ -25,7 +25,7 @@ public class SRGlobalLootModifierProvider extends GlobalLootModifierProvider {
         FlintExpansion.addGlobalLoot(this);
         Minecarts.addGlobalLoot(this);
         CoalFire.addGlobalLoot(this);
-        EnchantmentsFeature.addGlobalLoot(this);
+        Lapis.addGlobalLoot(this);
         RecallIdol.addGlobalLoot(this);
     }
 }